pyproject.toml
name = "ranch-pipeline"
version = "0.1.0"
description = "ICEAA 2026 — Ranch Self-Build Cost Engineering Pipeline (Workflow 3)"
authors = ["Rob Smale"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.0"
numpy = "^1.26"
openpyxl = "^3.1"
scipy = "^1.11"
matplotlib = "^3.8"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
